+++ cmd/hidden_lake/helpers/traffic/pkg/settings/settings.go
@@ -24,7 +24,7 @@ const (
 )
 
 const (
-	CDefaultStorageEnabled = false
+	CDefaultDatabaseEnabled = false
 )
 
 const (